requirements.txt
argparse
requests
pyyaml